moved Vector to its own module
This is because I'm going to make a Snake game, which requires the Vector class too.
